fixed a bug in `getDateRelative()` by which a page from yesterday can be labelled "today".
The bug in the function getDateRelative() results in a page created at 18:00 being output at 13:00 on the following day today as relative date, which is wrong.

I have modified getDateRelative() with the help of Mark so that the basis for calculating the time difference is the date and not the time (86400s (24h)) so now the date of the publication is checked against today's date and outputs the correct relative date.